Destructuring de Beneficios

+++ src/models/Beneficios.js
@@ -4,7 +4,7 @@ class Beneficios extends Model {
     static init(sequelize) {
         super.init({
             beneficio: DataTypes.STRING,
-            valor_beneficio: DataTypes.INTEGER,
+            valor_beneficio: DataTypes.FLOAT,
         }, {
             sequelize,//sequelize: connection
             // freezeTableName : true
